api/metadata: clean package (#926)
* clean api proto
* fix nil trace_id
