--- api/proto/kratos/api/annotations.proto
@@ -1,13 +1,0 @@
-syntax = "proto3";
-package kratos.api;
-
-option go_package = "github.com/go-kratos/kratos/v2/api/proto/kratos/api;api";
-option java_multiple_files = true;
-option java_package = "com.github.kratos.api";
-option objc_class_prefix = "KratosAPI";
-
-import "google/protobuf/descriptor.proto";
-
-extend google.protobuf.EnumOptions {
-	bool errors = 1108;
-}
