--- api/proto/kratos/api/metadata.proto
@@ -1,43 +1,0 @@
-syntax = "proto3";
-
-package kratos.api;
-
-import "google/protobuf/any.proto";
-import "google/protobuf/descriptor.proto";
-import "google/api/annotations.proto";
-
-option go_package = "github.com/go-kratos/kratos/v2/api/proto/kratos/api;api";
-option java_multiple_files = true;
-option java_package = "com.github.kratos.api";
-option objc_class_prefix = "KratosAPI";
-
-
-// Metadata is api defintion metadata service
-service Metadata {
-  // ListServices list the full name of all services
-  rpc ListServices (google.protobuf.Any) returns (ListServicesReply)  {
-      option (google.api.http) = {
-        get: "/services",
-      };
-  }
-
-  // GetServiceMeta get the full fileDescriptorSet of service
-  rpc GetServiceMeta (GetServiceMetaRequest) returns (GetServiceMetaReply)  {
-      option (google.api.http) = {
-        get: "/service/{name}/metadata",
-      };
-  }
-}
-
-message ListServicesReply {
-  repeated string services = 1;
-}
-
-message GetServiceMetaRequest {
-  string name = 1;
-}
-
-message GetServiceMetaReply {
-  google.protobuf.FileDescriptorSet protoSet = 1;
-}
-
