+++ examples/blog/api/blog/v1/error_reason.proto
@@ -1,0 +1,14 @@
+syntax = "proto3";
+
+package api.blog.v1;
+
+// 多语言特定包名，用于源代码引用
+option go_package = "github.com/go-kratos/kratos/examples/blog/api/v1;v1";
+option java_multiple_files = true;
+option java_package = "blog.v1.errors";
+option objc_class_prefix = "APIBlogErrors";
+
+enum ErrorReason {
+    TITLE_MISSING = 0;
+    CONTENTMISSING = 1;
+}
