--- examples/blog/api/blog/v1/errors/article.proto
@@ -1,18 +1,0 @@
-syntax = "proto3";
-
-package api.blog.v1.errors;
-
-import "kratos/api/annotations.proto";
-
-// 多语言特定包名，用于源代码引用
-option go_package = "github.com/go-kratos/kratos/examples/blog/api/v1/errors;errors";
-option java_multiple_files = true;
-option java_package = "blog.v1.errors";
-option objc_class_prefix = "APIBlogErrors";
-
-enum Article {
-    option (kratos.api.errors) = true;
-
-    TitleMissing = 0;
-    ContentMissing = 1;
-}
